+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Caching of successful verification (eng/recordflux/RecordFlux#1345)
+
 ## [0.11.0] - 2023-06-16
 
 ### Removed
@@ -319,6 +325,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2019-05-14
 
+[Unreleased]: https://github.com/AdaCore/RecordFlux/compare/v0.11.0...HEAD
 [0.11.0]: https://github.com/AdaCore/RecordFlux/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/AdaCore/RecordFlux/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/AdaCore/RecordFlux/compare/v0.9.0...v0.9.1